[file] 기대평 컴포넌트 분리
